+++ sdk/python/pulumi_gcp/cloudfunctions/function.py
@@ -110,7 +110,7 @@ class Function(pulumi.CustomResource):
         > **Warning:** As of November 1, 2019, newly created Functions are
         private-by-default and will require [appropriate IAM permissions](https://cloud.google.com/functions/docs/reference/iam/roles)
         to be invoked. See below examples for how to set up the appropriate permissions,
-        or view the [Cloud Functions IAM resources](https://www.terraform.io/docs/r/cloudfunctions_cloud_function_iam.html)
+        or view the [Cloud Functions IAM resources](https://www.terraform.io/docs/providers/google/r/cloudfunctions_cloud_function_iam.html)
         for Cloud Functions.
         
         :param str resource_name: The name of the resource.